refactor(workspace-store): reorganize type guards and improve imports
- Moved type guard functions for OpenAPI and AsyncAPI documents to a new helper file for better organization and clarity.
- Updated import statements across various files to utilize the new type guard location, enhancing maintainability.
- Added AsyncAPI document source to test data for improved testing coverage.
