+++ packages/api-reference/src/features/Search/hooks/useSearchIndex.ts
@@ -1,6 +1,6 @@
+import { isOpenApiDocument } from '@scalar/workspace-store/helpers/type-guards'
 import type { AsyncApiDocument } from '@scalar/workspace-store/schemas/asyncapi/asyncapi-document'
 import type { OpenApiDocument } from '@scalar/workspace-store/schemas/v3.1/strict/openapi-document'
-import { isOpenApiDocument } from '@scalar/workspace-store/schemas/workspace'
 import type { FuseResult } from 'fuse.js'
 import { computed, ref, toValue, watch } from 'vue'
