+++ packages/workspace-store/src/schemas/workspace.ts
@@ -1,7 +1,6 @@
 import { Type } from '@scalar/typebox'
 import { AVAILABLE_CLIENTS, type AvailableClients } from '@scalar/types/snippetz'
 
-import type { UnknownObject } from '@/helpers/general'
 import { compose } from '@/schemas/compose'
 import { extensions } from '@/schemas/extensions'
 import {
@@ -25,15 +24,6 @@ import {
 
 // Union type for API documents (OpenAPI or AsyncAPI)
 export type ApiDefinition = OpenApiDocument | AsyncApiDocument
-
-// Type guards
-export function isOpenApiDocument(doc: ApiDefinition | UnknownObject): doc is OpenApiDocument {
-  return 'openapi' in doc || 'swagger' in doc
-}
-
-export function isAsyncApiDocument(doc: ApiDefinition | UnknownObject): doc is AsyncApiDocument {
-  return 'asyncapi' in doc
-}
 
 export const WorkspaceDocumentMetaSchema = Type.Partial(
   Type.Object({
